# [dbnode] Decoder: fix handling of values requiring 64bit precision